New list type "FreezableList" allows locking an EventList from receiving events
This may cause minor issues as the EventList interface has changed
as a consequence of this change. See the EventList source for
details on this change and how it may impact you. The change
will only impact you if you implement EventList directly and do not
extend MutationList


Issue number:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.java.net/svn/glazedlists~svn/trunk@49 696c4f5c-992c-4387-df9e-f6c076136008