fixes by Kevin. He's an amazing coder man.
Kev's email:

I thought about the other bug a bunch and what I arrived at was kind of cool so I figured it worthy of sharing.

Calling the source list's methods (which are all on the SortedList of course) guarantees O(log(n)) runtime.  As it stood, the UniqueList methods each had a runtime of O(log(n)*log(n)).

What's kind of cool, is that the O(log(n)*log(n)) can be faster in some cases.  Since n can be sufficiently small, I decided to figure out exactly what the relation between size n of the compressed list and size x of the source list would have to be in order for the current implementation to be faster.

And here's the math behind it:

for a compressed list of size n:

Binary searching the compressed list is faster
iff
The source list is of size x where
2^((log(n))^2) < x <= 2^((log(n) + 1)^2)

with real numbers:
for a compressed view of size 8,
it is faster for a source list of size x where:
512 < x <= 65536

For this case, this shows that the compressed view must be strictly less than 1.5% of the source list size in order for the current implementation to be faster.  Clearly this case is sufficiently small that it works well as the worst case for the various methods.

Cool huh?  The fix is in the attached zip.

Cheers,
kev



Jesse's response:



First off, we have two size values:
N = size of sorted list
M = size of unique list, M <= N

So your optimization idea is as follows:
1. Binary search on UniqueList takes log(M) * log(M) time
2. indexof() on SortedList takes log(N) time
3. There exist cases such that log(M) * log(M) < log(N)
  For example, if M = 8
    then log(8) = 3.
    then log(8) * log(8) = 9.
  So if log(N) < 9, then binary search on unique list is faster.
    so then log(N) < 9
    or N < 2^9
    or N < 512

This is well thought out and very valid. Good work.

But I have to electronically kick you in the teeth now, unfortunately.
Here's the problem. I fear assumption #1 is incorrect:
Binary search on UniqueList takes log(M) * log(M) time
Actually, each stage in the unique list binary search performs a get() on the sorted list.
So total time is actually log(M) * log(M) * log(N) time.
And log(M) * log(M) * log(N) > log(N)

So we have a minimum bound of log(N) if we are interested in accessing the sorted list.




Issue number:  62
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.java.net/svn/glazedlists~svn/trunk@307 696c4f5c-992c-4387-df9e-f6c076136008