Issue number: 239 Fixed a documentation error in BarcodeIterator.
Optimized remove operations on IndexedTree so that they only make one traversal of a tree instead of two.

Optimized setComparator() on SortedList to not rebuild the unsorted tree which saves n traversals.  This also permitted me to change how the reorder map values are obtained which reduced 4n traversals to just n to build the reorder map.

Created a performance test to determine how much faster resorting by a different Comparator has become.  The result on my reduced Mozilla bug db (only 2204 issues) that I use for testing is a 2 second increase on 100 randomized comparator changes from 13.5 seconds to 11.5 seconds.


git-svn-id: https://svn.java.net/svn/glazedlists~svn/trunk@917 696c4f5c-992c-4387-df9e-f6c076136008