Bugfix GLAZEDLISTS-580.
Using a custom header renderer with JXTableSupport was difficult,
because JXTableSupport#install(..) always resetted the header renderer
to the result of "JTableHeader().getDefaultRenderer()".

This fix removes this behaviour, e.g. JXTableSupport doesn't replace the
defualt table header renderer anymore.
It's the responsibility of the API user to use a compatible header
renderer.