SeparatorList: Handling listChanges like __UX
Scenario: A unique element of a group is updated, and the remaining
elements of the group are deleted.

There were fixes in GLAZEDLISTS-522 and pull request 15 that deal with
this scenario for GroupingList and UniqueList, but they do not
sufficiently fix SeparatorList.

Fix: Add additional logic to handle edge cases introduced in
GLAZEDLISTS-522. This constraint is only applied when the deletes are
at the end of the list. Additionally, for the SeparatorList, we still
need to delete an element from the separator barcode when this happens.
So, we make an update call to the grouper client, rather than doing
nothing.

Additionally, I ported over many of the tests from pull request 15
to equivalent versions for SeparatorList.