SeparatorList: Separator position is updated prematurely
Scenario: listChanges contain update and insert/delete later on in the
same group.

When we’re processing the update, we prematurely shift the separator
for the next group, due the fix for bug 500. However, once we process
the insert or delete later on in that group, the separator ends up in
the wrong place.

Fix: Pass an additional boolean from the grouper to the grouper client,
which is only used by SeparatorList. Only perform this shift when this
variable is true.

We only need to perform this shift when an element is updated from the
left group to the right group, or vice versa.