SortedList produces listChanges that break SeparatorList
Scenario: Delete element 'A', the last element of Group 1. Move element
'B' from Group 3 to the front of Group 2.

I Would expect list changes with 'X+'. However, it ends up being '+X'.
This results in the separator in the wrong spot.

This happens due to the order items are sorted in SortedList. B is
evaluated prior to A, and is inserted into the list AFTER 'A'
(unsorted items are ignored in sorting logic).

The result is that when we try to remove the node for 'A' in the
separator barcode, we erroneously delete the separator between groups 1
and 2.

Fix: Reuse the shifting logic used for bug 500. Shift if the index of
the element we are deleting is a separator in the barcode.

The ideal fix would be to try to correct the list changes that
SortedList produces, but that requires digging deep into the internals
of SortedList / SimpleTree, could have unexpected side effects.