SeparatorList: Changes like __UXX, deletes are from the same group
Scenario: When deleting a non-unique it's possible that we end up
trying to update a separator at a position past the end of the list.

Fix: When performing a group update in the SeparatorList GrouperClient,
cap the groupIndex such that it points no further than the last
separator. This prevents the possibility of trying to update a
non-existent separator, which produces NPEs.