SeparatorList: Updates like __UXX, deletes from different groups
Scenario: listChanges contains an update on the unique element of a
group, and deletes the remaining elements of that group. Additionally,
the unique element (with no duplicates) of the following group is
deleted.

Both deleted elements have the same changeIndex, which is marked as the
'lastFakedUniqueChangeIndex'. While this is indeed the case for the
first deleted element, the second deleted eleemnt is truly unique. So,
for the first delete, we only want to update the group, while for the
second delete we should delete its group. However, since we end up
thinking that both deletes are from a faked unique change index, we
don't delete any groups.

Fix: Reset the lastFakedUniqueChangeIndex after it is processed for the
first time.