Fix for when updates occur at the boundaries between groups
There were a few problematic cases that arose when a new group is
created next to a group that is being updated. This is similar to some
of the fixes for the Grouper that we implemented recently. I believe
some of these cases may have impacted the GroupingList as well.

- needs test cleanup
- impacts GroupingList too (but haven't added a separate test,
it may not be possible w/ current structure)