Additional fixes to shifting logic
Some refinements needed to be made for when the separators should be
shifted.

1) We shift the separator if we detect that we are about to delete a
separator, when we should actually be deleting a white node. However,
in some cases we were actually attempting to shift the wrong separator,
which ended up having no action. Instead, we should be trying to shift
successive separators, until we will be deleting a white node.

2) Change the section of code beginning with the comment "special case
out the shift operation". This comment notes that if the fix poses a
problem, that we may want to change the grouper to fire a special flag
with the value true whenever the group joined is a RIGHT_GROUP. I did
indeed find a scenario where this caused problems, and added a test to
demonstrate it. So, I made the suggested change, and had the grouper
fire a flag called "joinRight".