Cause SWT tests to exit on Mac, per issue #636.
There is a Gradle bug (https://github.com/gradle/gradle/issues/864)
which prevents the "-XstartOnFirstThread" from working correctly and
thus prevents SWT unit tests from function on Mac. Since we're unable
to work around the Gradle issue, I've made this task exit with a
descriptive error when run on Mac. The instructions detail how to
suppress the SWT tests so that all other unit tests can be run.

Relates to: #636