Java 8 Collections method implementations for Abstract/BasicEL
I've tried to move all that seems reasonable into AbstractEventList so
it's picked up as widely as possible. I'm heavily leaning on the
removeIf(Predicate) method, so that will be one where performance is
important.

Also added a method to ListEventAssembler to allow testing of whether
an event is in progress or not. In addition to testing some good
behavior in AbstractEL, this will allow some future testing relating
to exceptions during event dispatching to subscribed lists.

Related: #644