removeIf implementation for FunctionList, work on FunctionListMap
FunctionList.removeIf is non-trvial because it jumps through some hoops
to try to fire a single ListEvent per remoteIf operation. To do so, the
source list must extends AbstractEventList.

FunctionListMap picks up a number of performance improvements, including
removing a usage of ListIterator, which is always a win. It also no
longer exposes its listChanged method.

Relates to: #644