AbstractEL: removeIf & batch addAll events, remove TransformedEL.clear
The clear implementation in TransformedList was no longer necessary due
to recent work in AbstractEventList.

Re-implement removeIf in AbstractEventList. It's almost exactly the same
as the default implementation in Collection but adds a batch event
wrapper.

Related to the above, a batch event wrapper was added to the addAll
method.

Relates to: #644